requirements.txt
Flask>=3.0
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
folium>=0.17
mysql-connector-python>=9.0.0
SQLAlchemy>=2.0
openpyxl>=3.1        # for reading .xlsx uploads
pytz>=2024.1         # used for "Live" Manila time
MarkupSafe>=2.1

# optional if you want local .env loading:
# python-dotenv>=1.0
